+++ packages/components/src/feedback/index.ts
@@ -1,0 +1,2 @@
+export { FeedbackModule } from './feedback.module';
+export { FeedbackComponent } from './feedback.component';
